+++ docs/CHANGELOG.md
@@ -7,7 +7,8 @@
 ### Improve
 
 * Update dependencies [#508](https://github.com/Riey/kime/issues/508)
-* Fix sebeolsik-3sin-p2 '"' character [#509](https://github.com/Riey/kime/pull/509)
+* Fix sebeolsik-3sin-p2 '"' character [#509](https://github.com/Riey/kime/issues/509)
+* Fix sebeolsik-391 "S-F" key [#509](https://github.com/Riey/kime/issues/521)
 
 ## 2.5.5